[refactor-minimaxes] PR comments
