+++ src/app/jscaip/tests/Rules.spec.ts
@@ -31,15 +31,6 @@ class AbstractRules extends Rules<P4Move, MyAbstractState> {
     }
 }
 
-describe('GameStatus', () => {
-    it('should be convertible to BoardValue', () => {
-        expect(GameStatus.ZERO_WON.toBoardValue().value).toBe(Player.ZERO.getVictoryValue());
-        expect(GameStatus.ONE_WON.toBoardValue().value).toBe(Player.ONE.getVictoryValue());
-        expect(GameStatus.DRAW.toBoardValue().value).toBe(0);
-        expect(GameStatus.ONGOING.toBoardValue().value).toBe(0);
-    });
-});
-
 describe('Rules', () => {
 
     let rules: AbstractRules;
